Introduce spotbugs

+++ bboxdb-server/src/main/java/org/bboxdb/distribution/allocator/MemoryUtilizationFreeAllocator.java
@@ -51,7 +51,7 @@ public class MemoryUtilizationFreeAllocator extends AbstractUtilizationAllocator
 			return 0;
 		}
 		
-		return distributedInstance.getMemory() / count;
+		return (double) distributedInstance.getMemory() / count;
 	}
 	
 	@Override
